refactor: 상수 폴더 별 관리
