+++ __tests__/constants/index.js
@@ -1,0 +1,7 @@
+export const TEST_CARS = ["car1", "car2", "car3"];
+
+export const TEST_DUPLICATED_CARS = ["car1", "car1", "car2"];
+
+export const TEST_NONEXISTENT_CARS = [];
+
+export const TEST_WHITESPACE_CARS = [" ", "   ", "    "];
