+++ docs/REQUIREMENTS.md
@@ -30,3 +30,4 @@
 - [x] 모든 메서드가 static -> 객체로 변경
 - [x] 역할 측면에서 for문 이외에 move를 결정하는 함수를 별도로 분리
 - [x] `randomValue`는 `Round`에 위임
+- [x] 상수 폴더 별 관리
