+++ src/constants/index.js
@@ -1,11 +1,3 @@
-export const TEST_CARS = ["car1", "car2", "car3"];
-
-export const TEST_DUPLICATED_CARS = ["car1", "car1", "car2"];
-
-export const TEST_NONEXISTENT_CARS = [];
-
-export const TEST_WHITESPACE_CARS = [" ", "   ", "    "];
-
 export const MOVE_THRESHOLD = 4;
 
 export const RANDOM_BOUND = 10;
